Fault-Detection-Module/requirements.txt
flask
kafka-python
tensorflow
grpcio
requests